# coursework/backend/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Krusty Crab CRM API
  description: |
    API для системы управления рестораном "Красти Крабс"
    
    Система предназначена для автоматизации бизнес-процессов ресторана:
    - Управление заказами и клиентами
    - Контроль меню и ингредиентов
    - Управление сотрудниками и рабочими сменами
    - Финансовый учет и отчётность
    - Программы лояльности
    - Мониторинг качества обслуживания
  version: 1.0.0
  contact:
    name: Krusty Crab Support
servers:
  - url: http://localhost:13228/api
    description: Local development server
...
paths:
...
  /orders/{orderId}:
    get:
      tags:
        - orders
      summary: Получить информацию о заказе
      description: Получение детальной информации о заказе (FR-003)
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Информация о заказе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - orders
      summary: Обновить статус заказа
      description: Обновление статуса заказа (FR-008, FR-010, FR-013)
      operationId: updateOrderStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Статус заказа обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /inventory/low-stock:
    get:
      tags:
        - inventory
      summary: Получить ингредиенты с низким остатком
      description: Получение списка ингредиентов с остатком ниже порога (FR-015)
      operationId: getLowStock
      security:
        - bearerAuth: []
      parameters:
        - name: thresholdFactor
          in: query
          description: Коэффициент порога
          schema:
            type: number
            format: double
            default: 1.0
      responses:
        '200':
          description: Список ингредиентов с низким остатком
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LowStockItem'
...
  /shifts:
    get:
      tags:
        - shifts
      summary: Получить список смен
      description: Получение списка всех смен (FR-011, FR-017)
      operationId: getAllShifts
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          description: Фильтр по дате
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Список смен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'

    post:
      tags:
        - shifts
      summary: Создать смену
      description: Создание новой смены (FR-017, только для управляющего)
      operationId: createShift
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftCreateRequest'
      responses:
        '201':
          description: Смена создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          $ref: '#/components/responses/BadRequest'
  /shifts/{shiftId}:
    get:
      tags:
        - shifts
      summary: Получить информацию о смене
      description: Получение информации о смене
      operationId: getShiftById
      security:
        - bearerAuth: []
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о смене
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - shifts
      summary: Обновить смену
      description: Обновление информации о смене (FR-017, только для управляющего)
      operationId: updateShift
      security:
        - bearerAuth: []
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftCreateRequest'
      responses:
        '200':
          description: Смена обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - shifts
      summary: Удалить смену
      description: Удаление смены (FR-017, только для управляющего)
      operationId: deleteShift
      security:
        - bearerAuth: []
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Смена удалена
        '404':
          $ref: '#/components/responses/NotFound'
  /shifts/{shiftId}/assign:
    post:
      tags:
        - shifts
      summary: Назначить сотрудника на смену
      description: Назначение сотрудника на смену (FR-017, только для управляющего)
      operationId: assignEmployeeToShift
      security:
        - bearerAuth: []
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - employeeId
              properties:
                employeeId:
                  type: integer
      responses:
        '201':
          description: Сотрудник назначен на смену
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeShift'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /shifts/assignments:
    get:
      tags:
        - shifts
      summary: Получить назначения сотрудников на смены
      description: Список назначений сотрудников на смены (для управляющего)
      operationId: getShiftAssignments
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          description: Фильтр по дате смены
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Назначения сотрудников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeShift'
  /shifts/assignments/{employeeShiftId}:
    delete:
      tags:
        - shifts
      summary: Снять сотрудника со смены
      description: Удаление назначения сотрудника на смену (FR-017, только для управляющего)
      operationId: removeEmployeeFromShift
      security:
        - bearerAuth: []
      parameters:
        - name: employeeShiftId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Назначение удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ========== KITCHEN ==========
  /kitchen/queue:
    get:
      tags:
        - kitchen
      summary: Получить очередь заказов для кухни
      description: Получение очереди заказов для кухни (FR-013)
      operationId: getKitchenQueue
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Очередь заказов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KitchenQueueItem'

  # ========== ANALYTICS ==========
  /analytics/sales:
    get:
      tags:
        - analytics
      summary: Получить сводку по продажам
      description: Получение сводки по продажам за период (FR-014, FR-020, для управляющего)
      operationId: getSalesSummary
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          description: Начало периода
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Конец периода
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Сводка по продажам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesSummary'
  /analytics/top-menu-items:
    get:
      tags:
        - analytics
      summary: Получить топ блюд
      description: Получение топ блюд по количеству и выручке (FR-012, FR-020, для управляющего)
      operationId: getTopMenuItems
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          description: Начало периода
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Конец периода
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Количество записей
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Топ блюд
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopMenuItem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен доступа. Получите токен через /auth/login/client или /auth/login/employee
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
      description: ID заказа
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Courier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicleInfo:
          type: string
        available:
          type: boolean
        busy:
          type: boolean
...
    Order:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        createdByEmployeeId:
          type: integer
          nullable: true
        acceptedByEmployeeId:
          type: integer
          nullable: true
        type:
          $ref: '#/components/schemas/OrderType'
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        preparingAt:
          type: string
          format: date-time
          nullable: true
        readyAt:
          type: string
          format: date-time
          nullable: true
        cookingDurationSeconds:
          type: integer
          minimum: 0
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalAmount:
          type: number
          format: decimal
        courierId:
          type: integer
          nullable: true
        courier:
          $ref: '#/components/schemas/Courier'
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
    OrderType:
      type: string
      enum:
        - dine_in
        - takeout
        - delivery
    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - preparing
        - ready
        - delivering
        - delivered
        - completed
        - cancelled
...
    PaymentMethod:
      type: string
      enum:
        - cash
        - card
        - online
...
    LowStockItem:
      type: object
      properties:
        ingredientId:
          type: integer
        name:
          type: string
        quantity:
          type: number
          format: decimal
        minThreshold:
          type: number
          format: decimal
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        shiftDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        note:
          type: string
    ShiftCreateRequest:
      type: object
      required:
        - shiftDate
        - startTime
        - endTime
      properties:
        shiftDate:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        note:
          type: string
    EmployeeShift:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        shiftId:
          type: integer
        status:
          type: string
    KitchenQueueItem:
      type: object
      properties:
        orderId:
          type: integer
        createdAt:
          type: string
          format: date-time
        status:
          type: string
        preparingAt:
          type: string
          format: date-time
          nullable: true
        readyAt:
          type: string
          format: date-time
          nullable: true
        cookingDurationSeconds:
          type: integer
          minimum: 0
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemInfo'
    OrderItemInfo:
      type: object
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        note:
          type: string
    SalesSummary:
      type: object
      properties:
        hasData:
          type: boolean
        fromTs:
          type: string
          format: date-time
        toTs:
          type: string
          format: date-time
        ordersCnt:
          type: integer
        revenue:
          type: number
          format: decimal
        avgTicket:
          type: number
          format: decimal
...
    TopMenuItem:
      type: object
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        revenue:
          type: number
          format: decimal
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string